Dashboard, Correzione Bug

+++ Hackathon_Team5_19_21/Data/ServizioCorsi.cs
@@ -16,21 +16,38 @@ namespace Hackathon_Team5_19_21.Data
             return await _db.Corsi.FindAsync(id);
         }
 
-        public async Task EliminaCorso(Corso c)
+        public async Task<bool> EliminaCorso(Corso c)
         {
-            _db.Corsi.Remove(c);
-            await SalvaCambiamenti();
+            bool result = false;
+            Modulo m = await _db.Moduli.FirstOrDefaultAsync(x => x.IdCorso == c.Id);
+            StudenteIscritto s = await _db.StudentiIscritti.FirstOrDefaultAsync(x => x.IdCorso == c.Id);
+
+            if (m == null && s == null)
+            {
+                _db.Corsi.Remove(c);
+                await _db.SaveChangesAsync();
+                result = true;
+            }
+            return result;
         }
 
         public async Task<List<Corso>> GetCorsi()
         {
-            return await _db.Corsi.OrderByDescending(x => x.AnnoInizio).ThenBy(x=>x.Id).ToListAsync();
+            return await _db.Corsi.OrderBy(x => x.Id).ToListAsync();
         }
 
-        public async Task SalvaCorso(Corso c)
+        public async Task<bool> SalvaCorso(Corso c)
         {
-            await _db.Corsi.AddAsync(c);
-            await SalvaCambiamenti();
+            bool result = false;
+            Citta citta = await _db.Citta.FindAsync(c.IdCitta);
+            PersonaFitstic p = await _db.PersonaleFitstic.FindAsync(c.IdOrganizzatore);
+            if (citta != null && p != null)
+            {
+                _db.Corsi.Add(c);
+                await _db.SaveChangesAsync();
+                result = true;
+            }
+            return result;
         }
         public async Task SalvaCambiamenti()
         {
@@ -39,7 +56,7 @@ namespace Hackathon_Team5_19_21.Data
 
         public async Task<List<Corso>> GetCorsiAttuali()
         {
-            return await _db.Corsi.Where(x=>x.AnnoFine>=DateTime.Now.Year).OrderByDescending(x => x.AnnoInizio).ThenBy(x => x.Id).ToListAsync();
+            return await _db.Corsi.Where(x => x.AnnoFine >= DateTime.Now.Year).OrderByDescending(x => x.AnnoInizio).ThenBy(x => x.Id).ToListAsync();
         }
     }
 }
